improve css handling
