+++ app/src/app.jsx
@@ -10,9 +10,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { Redirect, Link, Router, Switch, useLocation, useHistory } from "react-router-dom";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { setYoung } from "./redux/auth/actions";
-
 import { startReactDsfr } from "@codegouvfr/react-dsfr/spa";
-
 import { environment, maintenance } from "./config";
 import api, { initApi } from "./services/api";
 import { queryClient } from "./services/react-query";
@@ -20,7 +18,7 @@ import { YOUNG_STATUS } from "./utils";
 import { isFeatureEnabled, FEATURES_NAME } from "snu-lib";
 import { cohortsInit, getCohort } from "./utils/cohorts";
 
-import Loader from "./components/Loader";
+import PageLoader from "./components/PageLoader";
 import FallbackComponent from "./components/FallBackComponent";
 
 const AccountAlreadyExists = lazy(() => import("./scenes/account/AccountAlreadyExists"));
@@ -49,68 +47,63 @@ function App() {
   const dispatch = useDispatch();
   const history = useHistory();
 
-  async function fetchData() {
-    try {
-      const { ok, user, token } = await api.checkToken();
-      if (!ok) {
-        api.setToken(null);
-        dispatch(setYoung(null));
-      }
-      await cohortsInit();
-      if (token) api.setToken(token);
-      if (ok && user) {
-        dispatch(setYoung(user));
-        const cohort = await getCohort(user.cohort);
-
-        const isEmailValidationEnabled = isFeatureEnabled(FEATURES_NAME.EMAIL_VALIDATION, undefined, environment);
-        const forceEmailValidation =
-          isEmailValidationEnabled && user.status === YOUNG_STATUS.IN_PROGRESS && user.emailVerified === "false" && new Date() < new Date(cohort.inscriptionModificationEndDate);
-        if (forceEmailValidation) return history.push("/preinscription");
-      }
-    } catch (e) {
-      console.log(e);
-    } finally {
-      setLoading(false);
-    }
-  }
-
   useEffect(() => {
+    async function fetchData() {
+      try {
+        const { ok, user, token } = await api.checkToken();
+        if (!ok) {
+          api.setToken(null);
+          dispatch(setYoung(null));
+        }
+        await cohortsInit();
+        if (token) api.setToken(token);
+        if (ok && user) {
+          dispatch(setYoung(user));
+          const cohort = await getCohort(user.cohort);
+
+          const isEmailValidationEnabled = isFeatureEnabled(FEATURES_NAME.EMAIL_VALIDATION, undefined, environment);
+          const forceEmailValidation =
+            isEmailValidationEnabled && user.status === YOUNG_STATUS.IN_PROGRESS && user.emailVerified === "false" && new Date() < new Date(cohort.inscriptionModificationEndDate);
+          if (forceEmailValidation) return history.push("/preinscription");
+        }
+      } catch (e) {
+        console.log(e);
+      } finally {
+        setLoading(false);
+      }
+    }
     fetchData();
   }, []);
 
-  if (loading) return <Loader />;
+  if (loading) return <PageLoader />;
+
+  if (maintenance) return <Maintenance />;
 
   return (
     <Sentry.ErrorBoundary fallback={FallbackComponent}>
       <QueryClientProvider client={queryClient}>
         <Router history={sentryHistory}>
           <AutoScroll />
-          <Suspense fallback={<Loader />}>
-            {maintenance ? (
-              <Switch>
-                <SentryRoute path="/" component={Maintenance} />
-              </Switch>
-            ) : (
-              <Switch>
-                <PublicRoute path="/validate-contract/done" component={ContractDone} />
-                <PublicRoute path="/validate-contract" component={Contract} />
-                <PublicRoute path="/conditions-generales-utilisation" component={CGU} />
-                <PublicRoute path="/noneligible" component={NonEligible} />
-                <PublicRoute path="/representants-legaux" component={RepresentantsLegaux} />
-                <PublicRoute path="/je-rejoins-ma-classe-engagee" component={OnBoarding} />
-                <PublicRoute path="/je-suis-deja-inscrit" component={AccountAlreadyExists} />
-                <PublicRoute path="/public-besoin-d-aide" component={Contact} />
-                <PublicRoute path="/besoin-d-aide/ticket/:id" component={ViewMessage} />
-                <PublicRoute path="/besoin-d-aide" component={Contact} />
-                <PublicRoute path="/auth" component={Auth} />
-                <PublicRoute path="/public-engagements" component={AllEngagements} />
-                <PublicRoute path="/merci" component={Thanks} />
-                <PublicRoute path="/preinscription" component={PreInscription} />
-                <SecureRoute path="/inscription2023" component={Inscription2023} />
-                <SecureRoute path="/reinscription" component={ReInscription} />
-                <SecureRoute path="/" component={Espace} />
-              </Switch>
-            )}
+          <Suspense fallback={<PageLoader />}>
+            <Switch>
+              <PublicRoute path="/validate-contract/done" component={ContractDone} />
+              <PublicRoute path="/validate-contract" component={Contract} />
+              <PublicRoute path="/conditions-generales-utilisation" component={CGU} />
+              <PublicRoute path="/noneligible" component={NonEligible} />
+              <PublicRoute path="/representants-legaux" component={RepresentantsLegaux} />
+              <PublicRoute path="/je-rejoins-ma-classe-engagee" component={OnBoarding} />
+              <PublicRoute path="/je-suis-deja-inscrit" component={AccountAlreadyExists} />
+              <PublicRoute path="/public-besoin-d-aide" component={Contact} />
+              <PublicRoute path="/besoin-d-aide/ticket/:id" component={ViewMessage} />
+              <PublicRoute path="/besoin-d-aide" component={Contact} />
+              <PublicRoute path="/auth" component={Auth} />
+              <PublicRoute path="/public-engagements" component={AllEngagements} />
+              <PublicRoute path="/merci" component={Thanks} />
+              <PublicRoute path="/preinscription" component={PreInscription} />
+              <SecureRoute path="/inscription2023" component={Inscription2023} />
+              <SecureRoute path="/reinscription" component={ReInscription} />
+              <SecureRoute path="/" component={Espace} />
+            </Switch>
           </Suspense>
         </Router>
       </QueryClientProvider>
