+++ app/src/components/PageLoader.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import Loader from "./Loader";
+
+export default function PageLoader() {
+  return (
+    <div className="min-h-screen flex items-center justify-center">
+      <Loader />
+    </div>
+  );
+}
