+++ app/src/scenes/inscription2023/index.jsx
@@ -24,7 +24,6 @@ import { getCohort } from "@/utils/cohorts";
 import useAuth from "@/services/useAuth";
 import Help from "./components/Help";
 import Stepper from "@/components/dsfr/ui/Stepper";
-import useDocumentCss from "@/hooks/useDocumentCss";
 
 function renderStep(step) {
   if (step === STEPS.COORDONNEES) return <StepCoordonnees />;
@@ -104,8 +103,6 @@ const StepCorrection = () => {
 };
 
 export default function Index() {
-  useDocumentCss(["/dsfr/utility/icons/icons.min.css", "/dsfr/dsfr.min.css"]);
-
   const young = useSelector((state) => state.Auth.young);
   const { isCLE } = useAuth();
   const cohort = getCohort(young.cohort);
